RF: Move get_current_lr/momentum into optim.utils
