+++ echofilter/optim/__init__.py
@@ -1,3 +1,4 @@
 from . import criterions
 from . import meters
 from . import schedulers
+from . import utils
